Custom Fake Time System Update
Made the animator variable in the player controller public.

Created a new power called Warp2 based on Warp1 using the new time system.

Created new scriptable object called Information in namespace SpectralDaze.ScriptableObjects

Created a new class called Information in SpectralDaze.Time to hold information for the scriptable objects.

Created a instance of the scriptable object for time and assigned it to the Level03_Damien player along with the new power.